+++ opencascade-hs/src/OpenCascade/GCPnts/Types.hs
@@ -1,0 +1,6 @@
+{-# LANGUAGE EmptyDataDecls #-}
+module OpenCascade.GCPnts.Types
+( AbscissaPoint
+) where
+
+data AbscissaPoint