+++ scripts/regenerate-images.sh
@@ -10,6 +10,7 @@ stack exec waterfall-cad-examples --  --bound --dark-mode-svg images/bounding-bo
 stack exec waterfall-cad-examples --  --loft --dark-mode-svg images/loft.svg
 stack exec waterfall-cad-examples --  --2d-booleans --dark-mode-svg images/2d-booleans.svg --height 200
 stack exec waterfall-cad-examples --  --platonic-solids --dark-mode-svg images/platonic.svg
+stack exec waterfall-cad-examples --  --take-path-fraction --dark-mode-svg images/takePathFraction.svg
 
 stack exec waterfall-cad-examples --  --csg --glb images/models/csg.glb
 stack exec waterfall-cad-examples --  --gear --glb images/models/gear.glb
@@ -23,3 +24,4 @@ stack exec waterfall-cad-examples --  --fillet --glb images/models/fillet.glb
 stack exec waterfall-cad-examples --  --prism --glb images/models/prism.glb
 stack exec waterfall-cad-examples --  --2d-booleans --glb images/models/2d-booleans.glb
 stack exec waterfall-cad-examples --  --platonic-solids --glb images/models/platonic.glb
+stack exec waterfall-cad-examples --  --take-path-fraction --glb images/models/takePathFraction.glb