+++ waterfall-cad/src/Waterfall/TwoD/Path2D.hs
@@ -27,7 +27,9 @@ module Waterfall.TwoD.Path2D
 , reversePath2D
 , splice2D
 , splitPath2D
-) where 
+, pathLength2D
+, takePathFraction2D
+) where
 
 import Waterfall.TwoD.Internal.Path2D (Path2D(..))
 import Waterfall.TwoD.Transforms (rotate2D)
@@ -159,3 +161,11 @@ splice2D = splice
 -- | `splitPath` with the type fixed to `Path2D`
 splitPath2D :: Path2D -> [Path2D]
 splitPath2D = splitPath
+
+-- | `pathLength` with the type fixed to `Path`
+pathLength2D :: Path2D -> Double
+pathLength2D = pathLength
+
+-- | `takePathFraction` with the type fixed to `Path2D`
+takePathFraction2D :: Double -> Path2D -> Path2D
+takePathFraction2D = takePathFraction