+++ waterfall-cad/CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to the
 
 ## 0.6.1.0
 
+- Added `Waterfall.Path.Common.pathLength`, `Waterfall.Path.pathLength3D` and `Waterfall.TwoD.Path2D.pathLength2D`
+- Added `Waterfall.Path.Common.takePathFraction`, `Waterfall.Path.takePathFraction3D` and `Waterfall.TwoD.Path2D.takePathFraction2D`
+
 ### Added 
 
 - Added `Waterfall.Solids.torus`